+++ src/main/java/de/com/fdm/microsub/consumer/MicrosubConsumerServiceImpl.java
@@ -7,8 +7,8 @@ import de.com.fdm.grpc.microsub.lib.EventsubMessage;
 import de.com.fdm.db.data.MicroSub;
 import de.com.fdm.db.repositories.MicroSubRepository;
 import de.com.fdm.grpc.microsub.lib.Type;
-import de.com.fdm.tmi.OutboundMessage;
-import de.com.fdm.tmi.TmiService;
+import de.com.fdm.twitch.tmi.OutboundMessage;
+import de.com.fdm.twitch.tmi.TmiService;
 import io.grpc.Context;
 import io.grpc.stub.StreamObserver;
 import net.devh.boot.grpc.server.service.GrpcService;
